It was created the initial structure of the front and back end.
The frontend was just started with the vue create and vuetify, nothing more.

The backend was created using express and DDD. It was made the basic structure of files, database configuration with typeorm, the creation of the table students in the database, configuration of project patterns (editorconfig, prettier and eslint) and configuration of typescript. For now it doesn't has any feature working, just the basics of th project.